moved files to components folders
