+++ src/components/About.js
@@ -1,5 +1,5 @@
 import React from "react"
-import './App.css';
+import '../App.css';
 
 function About() {
   return (
