+++ src/components/Nav.js
@@ -1,7 +1,7 @@
 import React from "react"
-import './App.css';
+import '../App.css';
 import {Link} from 'react-router-dom';
-import Logo from './logo.png'
+import Logo from '../logo.png'
 
 function Nav() {
   return (
